refactor(types): extract shared agent↔web wire-protocol types
InventoryContainer, ContainerState, and AgentContainerEvent were
duplicated between agent/ and src/. Move them to
agent/src/types/protocol.ts (agent owns the wire format) and
re-export from the web side via the existing @homelab-manager/agent
path alias. Rename agent's ContainerEventMessage to AgentContainerEvent
for consistency with web naming.
